+++ src/Components/Input.js
@@ -102,8 +102,8 @@ const Input = props => {
 }
 
 Input.propTypes = {
-    id: PropTypes.string,
-    name: PropTypes.string,
+    id: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
+    name: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
     label: PropTypes.string,
     action: PropTypes.object,
     helpText: PropTypes.oneOfType([PropTypes.string, PropTypes.object]),